back up training scripts, etc

+++ asr_egs/wsj/steps/train_ctc_parallel_h.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
-
+{
 # Copyright 2015  Yajie Miao    (Carnegie Mellon University)
+#           2015  Hang Su
 # Apache 2.0
 
 # This script trains acoustic models based on CTC and using SGD. 
@@ -39,18 +40,23 @@ sort_by_len=true         # whether to sort the utterances by their lengths
 
 norm_vars=true           # whether to apply variance normalization when we do cmn
 add_deltas=true          # whether to add deltas
-copy_feats=true          # whether to copy features into a local dir (on the GPU machine)
-feats_tmpdir=            # the tmp dir to save the copied features, when copy_feats=true
 
 # status of learning rate schedule; useful when training is resumed from a break point
-cvacc=0
+cvacc=-1
 halving=0
+
+# Multi-GPU training
+nj=1
+utts_per_avg=700
+
+clean_up=true
 
 ## End configuration section
 
 echo "$0 $@"  # Print the command line for logging
 
-[ -f path.sh ] && . ./path.sh; 
+[ -f ./path.sh ] && . ./path.sh; 
+[ -f ./cmd.sh ] && . ./cmd.sh; 
 
 . utils/parse_options.sh || exit 1;
 
@@ -67,7 +73,7 @@ dir=$3
 mkdir -p $dir/log $dir/nnet
 
 for f in $data_tr/feats.scp $data_cv/feats.scp $dir/labels.tr.gz $dir/labels.cv.gz $dir/nnet.proto; do
-  [ ! -f $f ] && echo "decode.sh: no such file $f" && exit 1;
+  [ ! -f $f ] && echo "train_ctc_parallel.sh: no such file $f" && exit 1;
 done
 
 ## Read the training status for resuming
@@ -80,28 +86,19 @@ done
 echo $norm_vars > $dir/norm_vars  # output feature configs which will be used in decoding
 echo $add_deltas > $dir/add_deltas
 
-if $sort_by_len; then
-  feat-to-len scp:$data_tr/feats.scp ark,t:- | awk '{print $2}' > $dir/len.tmp || exit 1;
-  paste -d " " $data_tr/feats.scp $dir/len.tmp | sort -k3 -n - | awk '{print $1 " " $2}' > $dir/train.scp || exit 1;
-  feat-to-len scp:$data_cv/feats.scp ark,t:- | awk '{print $2}' > $dir/len.tmp || exit 1;
-  paste -d " " $data_cv/feats.scp $dir/len.tmp | sort -k3 -n - | awk '{print $1 " " $2}' > $dir/cv.scp || exit 1;
-  rm -f $dir/len.tmp
-else
-  cat $data_tr/feats.scp | utils/shuffle_list.pl --srand ${seed:-777} > $dir/train.scp
-  cat $data_cv/feats.scp | utils/shuffle_list.pl --srand ${seed:-777} > $dir/cv.scp
-fi
+echo "Preparing train and cv features"
+tmpdir=$dir/feats; 
+[ -d $tmpdir ] || mkdir -p $tmpdir
+[ $clean_up == true ] && trap "echo \"Removing features tmpdir $tmpdir @ $(hostname)\"; ls $tmpdir; rm -r $tmpdir" EXIT
+utils/prep_scps.sh --nj $nj --cmd "$train_cmd" ${seed:+ --seed=$seed} --clean-up $clean_up \
+  $data_tr/feats.scp $data_cv/feats.scp $num_sequence $frame_num_limit $tmpdir $dir || exit 1;
 
-feats_tr="ark,s,cs:apply-cmvn --norm-vars=$norm_vars --utt2spk=ark:$data_tr/utt2spk scp:$data_tr/cmvn.scp scp:$dir/train.scp ark:- |"
-feats_cv="ark,s,cs:apply-cmvn --norm-vars=$norm_vars --utt2spk=ark:$data_cv/utt2spk scp:$data_cv/cmvn.scp scp:$dir/cv.scp ark:- |"
+feats_tr="ark,s,cs:apply-cmvn --norm-vars=$norm_vars --utt2spk=ark:$data_tr/utt2spk scp:$data_tr/cmvn.scp scp:$dir/feats_tr.JOB.scp ark:- |"
+feats_cv="ark,s,cs:apply-cmvn --norm-vars=$norm_vars --utt2spk=ark:$data_cv/utt2spk scp:$data_cv/cmvn.scp scp:$dir/feats_cv.JOB.scp ark:- |"
 
-# Save the features to a local dir on the GPU machine. On Linux, this usually points to /tmp
-if $copy_feats; then
-  tmpdir=$(mktemp -d $feats_tmpdir);
-  copy-feats "$feats_tr" ark,scp:$tmpdir/train.ark,$dir/train_local.scp || exit 1;
-  copy-feats "$feats_cv" ark,scp:$tmpdir/cv.ark,$dir/cv_local.scp || exit 1;
-  feats_tr="ark,s,cs:copy-feats scp:$dir/train_local.scp ark:- |"
-  feats_cv="ark,s,cs:copy-feats scp:$dir/cv_local.scp ark:- |"
-  trap "echo \"Removing features tmpdir $tmpdir @ $(hostname)\"; ls $tmpdir; rm -r $tmpdir" EXIT
+if [ $nj -eq 1 ]; then
+  feats_tr=$(echo $feats_tr | sed 's#JOB#1#')
+  feats_tr=$(echo $feats_cv | sed 's#JOB#1#')
 fi
 
 if $add_deltas; then
@@ -133,33 +130,52 @@ for iter in $(seq $start_epoch_num $max_iters); do
     echo -n "EPOCH $iter RUNNING ... "
 
     # train
-    $train_tool --report-step=$report_step --num-sequence=$num_sequence --frame-limit=$frame_num_limit \
+    if [ -z "$nj" ]; then
+      $train_tool --report-step=$report_step --num-sequence=$num_sequence --frame-limit=$frame_num_limit \
         --learn-rate=$learn_rate --momentum=$momentum \
         --verbose=$verbose \
         "$feats_tr" "$labels_tr" $dir/nnet/nnet.iter$[iter-1] $dir/nnet/nnet.iter${iter} \
         >& $dir/log/tr.iter$iter.log || exit 1;
+      tracc=$(cat $dir/log/tr.iter${iter}.log | grep "TOKEN_ACCURACY" | tail -n 1 | awk '{ acc=$3; gsub("%","",acc); print acc; }')
+    else
+      $cuda_cmd JOB=1:$nj $dir/log/tr.iter$iter.JOB.log \
+        $train_tool --report-step=$report_step --num-sequence=$num_sequence --frame-limit=$frame_num_limit \
+        --learn-rate=$learn_rate --momentum=$momentum --num-jobs=$nj --job-id=JOB \
+        --verbose=$verbose \
+        ${utts_per_avg:+ --utts-per-avg=$utts_per_avg} \
+        "$feats_tr" "$labels_tr" $dir/nnet/nnet.iter$[iter-1] $dir/nnet/nnet.iter${iter} >& $dir/log/tr.iter$iter.log || exit 1
+      tracc=$(cat $dir/log/tr.iter${iter}.1.log | grep "TOTAL TOKEN_ACCURACY" | tail -n 1 | awk '{ acc=$(NF-1); gsub("%","",acc); print acc; }')
+    fi
 
+
+    echo -n "lrate $(printf "%.6g" $learn_rate), TRAIN ACCURACY $(printf "%.4f" $tracc)%, "
     end_time=`date | awk '{print $6 "-" $2 "-" $3 " " $4}'`
     echo -n "ENDS [$end_time]: "
 
-    tracc=$(cat $dir/log/tr.iter${iter}.log | grep "TOKEN_ACCURACY" | tail -n 1 | awk '{ acc=$3; gsub("%","",acc); print acc; }')
-    echo -n "lrate $(printf "%.6g" $learn_rate), TRAIN ACCURACY $(printf "%.4f" $tracc)%, "
-
     # validation
-    $train_tool --report-step=$report_step --num-sequence=$valid_num_sequence --frame-limit=$frame_num_limit \
+    if [ -z "$nj" ]; then
+      $train_tool --report-step=$report_step --num-sequence=$valid_num_sequence --frame-limit=$frame_num_limit \
         --cross-validate=true \
         --learn-rate=$learn_rate \
-        --momentum=$momentum \
         --verbose=$verbose \
         "$feats_cv" "$labels_cv" $dir/nnet/nnet.iter${iter} \
         >& $dir/log/cv.iter$iter.log || exit 1;
+      cvacc=$(cat $dir/log/cv.iter${iter}.log | grep "TOKEN_ACCURACY" | tail -n 1 | awk '{ acc=$3; gsub("%","",acc); print acc; }')
+    else
+      $cuda_cmd JOB=1:$nj $dir/log/cv.iter$iter.JOB.log \
+        $train_tool --report-step=$report_step --num-sequence=$valid_num_sequence --frame-limit=$frame_num_limit \
+        --cross-validate=true --num-jobs=$nj --job-id=JOB \
+        --learn-rate=$learn_rate \
+        --verbose=$verbose \
+        "$feats_cv" "$labels_cv" $dir/nnet/nnet.iter${iter} >& $dir/log/cv.iter$iter.log || exit 1;
+      cvacc=$(cat $dir/log/cv.iter${iter}.1.log | grep "TOTAL TOKEN_ACCURACY" | tail -n 1 | awk '{ acc=$(NF-1); gsub("%","",acc); print acc; }')
+    fi
 
-    cvacc=$(cat $dir/log/cv.iter${iter}.log | grep "TOKEN_ACCURACY" | tail -n 1 | awk '{ acc=$3; gsub("%","",acc); print acc; }')
     echo "VALID ACCURACY $(printf "%.4f" $cvacc)%"
 
     # stopping criterion
     rel_impr=$(bc <<< "($cvacc-$cvacc_prev)")
-    if [ 1 == $halving -a 1 == $(bc <<< "$rel_impr < $end_halving_inc") ]; then
+    if [[ 1 == "$halving" && 1 == $(bc <<< "$rel_impr < $end_halving_inc") ]]; then
       if [[ "$min_iters" != "" ]]; then
         if [ $min_iters -gt $iter ]; then
           echo we were supposed to finish, but we continue as min_iters : $min_iters
@@ -192,3 +208,4 @@ done
 format-to-nonparallel $dir/nnet/nnet.iter${iter} $dir/final.nnet >& $dir/log/model_to_nonparal.log || exit 1;
 
 echo "Training succeeded. The final model $dir/final.nnet"
+}
